Update after peer review #83
With this fix, for css, headjs and js resources a unique indetifier is
calculated with the fragment/component name and the relative path. That
unique identifier is stored and looked for when resolving resources, hence
prevent resolving resource helpers which are already resolved.

For tests, a method to create a mock RenderingFlowTracker is introduced. A
new test is intorduced to test duplicate resources in the final html. The
existing tests for css, headjs and js are also modified to use the newly
introdued mock RenderingFlowTracker.